requirements.txt
dj-database-url==0.5.0
cmake==3.18.4.post1
tensorflow-cpu
keras
Django==3.1.3
gunicorn==20.0.4
whitenoise==5.2.0
django-heroku==0.3.1
opencv-python-headless==4.5.1.48
Pillow==8.1.2
imutils==0.5.4
psycopg2==2.8.6
dropbox==11.5.0
dlib==19.21.1
